# Improve public API resource search route performance.