--- a/src/Aquifer.Public.API/Services/ServiceRegistry.cs
+++ b/src/Aquifer.Public.API/Services/ServiceRegistry.cs
@@ -1,0 +1,13 @@
+﻿using Aquifer.Common.Services.Caching;
+
+namespace Aquifer.Public.API.Services;
+
+public static class ServiceRegistry
+{
+    public static IServiceCollection AddCachingServices(this IServiceCollection services)
+    {
+        services.AddScoped<ICachingLanguageService, CachingLanguageService>();
+
+        return services;
+    }
+}