This improves performance of basic query searching by a factor of 2-3.  Most of the improvement is from changing the ordering but removing the language JOIN also shaved off some time.  See comments for details on both of those.  See also https://biblionexusworkspace.slack.com/archives/C05D8BE3KUY/p1728062517872939 for some additional ideas for future improvements.

Some data from QA DB profiling.  Searching for "Christ" (https://qa.api.aquifer.bible/resources/search?languageCode=eng&query=christ&offset=0&limit=100&api-key={apiKey}) takes the following duration:
* Getting the total: 1.012 seconds
* Getting the first 100 results (there are only 31): 3.782 seconds

With these changes applied:
* Getting the total: 0.426 seconds
* Getting the first 100 results (there are only 31): 1.700 seconds